chore(js,react): move options to be first class citizens on the inbox component
