+++ packages/js/src/ui/components/primitives/Tooltip/TooltipContent.tsx
@@ -1,6 +1,7 @@
 import { JSX, onCleanup, onMount, Show, splitProps } from 'solid-js';
 import { Portal } from 'solid-js/web';
-import { AppearanceKey, useFocusManager } from '../../../context';
+import { useFocusManager } from '../../../context';
+import type { AppearanceKey } from '../../../types';
 import { useStyle } from '../../../helpers';
 import { Root } from '../../elements';
 import { useTooltip } from './TooltipRoot';
