+++ packages/js/src/ui/config/index.ts
@@ -1,1 +1,3 @@
-export * from './defaultAppearance';
+export * from './defaultVariables';
+export * from './defaultLocalization';
+export * from './appearanceKeys';
