+++ packages/js/src/ui/context/AppearanceContext.tsx
@@ -11,159 +11,7 @@ import {
 import { createStore } from 'solid-js/store';
 import { defaultVariables } from '../config';
 import { parseElements, parseVariables } from '../helpers';
-
-export type CSSProperties = {
-  [key: string]: string | number;
-};
-
-export type ElementStyles = string | CSSProperties;
-
-/*
- * The double underscore signals that entire key extends the right part of the key
- * i.e. foo__bar means that foo_bar is an extension of bar. Both keys will be applied when foo_bar is used
- * meaning you would have `bar foo__bar` in the dom
- */
-export const appearanceKeys = [
-  //Primitives
-  'button',
-
-  'popoverContent',
-  'popoverTrigger',
-
-  'dropdownContent',
-  'dropdownTrigger',
-  'dropdownItem',
-  'dropdownItemLabel',
-  'dropdownItemLabelContainer',
-  'dropdownItemLeftIcon',
-  'dropdownItemRightIcon',
-
-  'tooltipContent',
-  'tooltipTrigger',
-
-  'back__button',
-
-  'skeletonText',
-  'skeletonAvatar',
-  'tabsRoot',
-  'tabsList',
-  'tabsContent',
-  'tabsTrigger',
-  'dots',
-
-  //General
-  'root',
-  'bellIcon',
-  'bellContainer',
-  'bellDot',
-  'preferences__button',
-  'preferencesContainer',
-  'inboxHeader',
-  'loading',
-
-  //Inbox
-  'inbox__popoverTrigger',
-  'inbox__popoverContent',
-
-  //Notifications
-  'notificationList',
-  'notificationListEmptyNoticeContainer',
-  'notificationListEmptyNotice',
-  'notificationListEmptyNoticeIcon',
-  'notificationListNewNotificationsNoticeContainer',
-  'notificationListNewNotificationsNotice__button',
-
-  'notification',
-  'notificationDot',
-  'notificationSubject',
-  'notificationBody',
-  'notificationBodyContainer',
-  'notificationImage',
-  'notificationDate',
-  'notificationDefaultActions',
-  'notificationCustomActions',
-  'notificationPrimaryAction__button',
-  'notificationSecondaryAction__button',
-  'notificationRead__button',
-  'notificationUnread__button',
-  'notificationArchive__button',
-  'notificationUnarchive__button',
-
-  // Notifications tabs
-  'notificationsTabs__tabsRoot',
-  'notificationsTabs__tabsList',
-  'notificationsTabs__tabsContent',
-  'notificationsTabs__tabsTrigger',
-  'notificationsTabsTriggerLabel',
-  'notificationsTabsTriggerCount',
-
-  //Inbox status
-  'inboxStatus__title',
-  'inboxStatus__dropdownTrigger',
-  'inboxStatus__dropdownContent',
-  'inboxStatus__dropdownItem',
-  'inboxStatus__dropdownItemLabel',
-  'inboxStatus__dropdownItemLabelContainer',
-  'inboxStatus__dropdownItemLeftIcon',
-  'inboxStatus__dropdownItemRightIcon',
-
-  // More actions
-  'moreActionsContainer',
-  'moreActions__dropdownTrigger',
-  'moreActions__dropdownContent',
-  'moreActions__dropdownItem',
-  'moreActions__dropdownItemLabel',
-  'moreActions__dropdownItemLeftIcon',
-
-  // More tabs
-  'moreTabs__button',
-  'moreTabs__dots',
-  'moreTabs__dropdownTrigger',
-  'moreTabs__dropdownContent',
-  'moreTabs__dropdownItem',
-  'moreTabs__dropdownItemLabel',
-  'moreTabs__dropdownItemRightIcon',
-
-  //workflow
-  'workflowContainer',
-  'workflowLabel',
-  'workflowLabelContainer',
-
-  // channel
-  'channelContainer',
-  'channelsContainer',
-  'channelLabel',
-  'channelLabelContainer',
-  'channelDescription',
-  'channelSwitchContainer',
-  'channelSwitch',
-  'channelSwitchThumb',
-
-  //Preferences Header
-  'preferencesHeader',
-  'preferencesHeader__back__button',
-  'preferencesHeader__title',
-
-  //Preferences Loading
-  'preferencesLoadingContainer',
-] as const;
-
-export type Variables = {
-  colorBackground?: string;
-  colorForeground?: string;
-  colorPrimary?: string;
-  colorPrimaryForeground?: string;
-  colorSecondary?: string;
-  colorSecondaryForeground?: string;
-  colorCounter?: string;
-  colorCounterForeground?: string;
-  colorNeutral?: string;
-  fontSize?: string;
-  borderRadius?: string;
-};
-
-export type AppearanceKey = (typeof appearanceKeys)[number];
-export type Elements = Partial<Record<AppearanceKey, ElementStyles>>;
+import type { Appearance, Elements, Variables } from '../types';
 
 type AppearanceContextType = {
   variables?: Variables;
@@ -173,9 +21,6 @@ type AppearanceContextType = {
 };
 
 const AppearanceContext = createContext<AppearanceContextType | undefined>(undefined);
-
-export type Theme = Pick<AppearanceContextType, 'elements' | 'variables'>;
-export type Appearance = Theme & { baseTheme?: Theme | Theme[] };
 
 type AppearanceProviderProps = ParentProps & { appearance?: Appearance } & { id: string };
