+++ packages/js/src/ui/index.ts
@@ -2,4 +2,4 @@ export { NovuUI } from './novuUI';
 export type { InboxProps } from './components';
 export type { BaseNovuUIOptions, NovuUIOptions } from './novuUI';
 export type { Notification } from '../notifications';
-export type { Tab, NotificationClickHandler, NotificationActionClickHandler } from './types';
+export * from './types';
