+++ playground/nextjs/src/pages/notifications/index.tsx
@@ -7,7 +7,7 @@ export default function Home() {
     <>
       <Title title="Notifications Component" />
       <div className="w-96 h-96 overflow-y-auto">
-        <Inbox options={novuConfig}>
+        <Inbox {...novuConfig}>
           <Notifications />
         </Inbox>
       </div>
